third_party/blink/web_tests/wpt_internal/webmcp/css/tool-active-pseudo-parsing.html

<!DOCTYPE html>
<title>Parsing tests for :tool-form-active and :tool-submit-active</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test_valid_selector(":tool-form-active");
  test_valid_selector(":tool-submit-active");
  test_valid_selector(":tool-form-active:tool-submit-active");
  test_valid_selector("::part(my-part):tool-form-active");
  test_valid_selector("::part(my-part):tool-submit-active");
  test_valid_selector(":is(:tool-form-active, :tool-submit-active)");
  test_valid_selector(":has(:tool-form-active)");
  test_valid_selector(":has(:tool-submit-active)");

  test_invalid_selector("::before:tool-form-active");
  test_invalid_selector("::before:tool-submit-active");
  test_invalid_selector("::selection:tool-form-active");
  test_invalid_selector("::selection:tool-submit-active");
</script>
